segger_rtt: re-org modules and clarify not-a-chip
